refactor: 닉네임 중복체크 위치 수정
