+++ apps/api/src/main/java/kr/flint/api/user/controller/spec/UserControllerDocs.java
@@ -1,0 +1,28 @@
+package kr.flint.api.user.controller.spec;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import kr.flint.shared.dto.response.SuccessResponse;
+import kr.flint.user.dto.response.NicknameCheckResponse;
+import org.springframework.http.ResponseEntity;
+
+@Tag(name = "User", description = "사용자 API")
+public interface UserControllerDocs {
+
+    @Operation(
+            summary = "닉네임 중복 체크",
+            description = "닉네임 사용 가능 여부를 확인합니다."
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "닉네임 체크 성공"
+            )
+    })
+    ResponseEntity<SuccessResponse<NicknameCheckResponse>> checkNickname(
+            @Parameter(description = "확인할 닉네임", example = "my_nickname") String nickname
+    );
+}
