--- modules/auth/src/main/java/kr/flint/auth/dto/response/NicknameCheckResponse.java
@@ -1,9 +1,0 @@
-package kr.flint.auth.dto.response;
-
-public record NicknameCheckResponse(
-        boolean available
-) {
-    public static NicknameCheckResponse of(boolean available) {
-        return new NicknameCheckResponse(available);
-    }
-}
